Good Auton for moving from Landfill to Auto (might be better to turn the opposite direction but that needs to be tested)

+++ src/org/usfirst/frc/team5407/robot/commands/ClawGrab.java
@@ -6,7 +6,7 @@ import edu.wpi.first.wpilibj.command.Command;
 
 public class ClawGrab extends Command{
 
-	public ClawGrab(int i) {
+	public ClawGrab(double i) {
 		requires(Robot.claw);
 		setTimeout(i);
 	}
